refactor: rename files and move package from domain to collection
