+++ src/main/java/lotto/collections/LottoTickets.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-import lotto.domain.LottoResult;
 import lotto.domain.LottoTicket;
 
 public class LottoTickets {
@@ -22,7 +21,7 @@ public class LottoTickets {
     return lottoTickets.size();
   }
 
-  public LottoResult calculateLottoResult(WinningNumbers numbers) {
+  public LottoResult createLottoResult(WinningNumbers numbers) {
     Map<Integer, Integer> lottoStatistics = lottoTickets.stream()
         .map(ticket -> ticket.getMatchCounts(numbers))
         .collect(Collectors
